extract `dynamic` module
